+++ src/aux-server/aux-web/shared/FileHelper.ts
@@ -31,6 +31,7 @@ import { Subject, Observable } from 'rxjs';
  * that help manipulate files.
  */
 export class FileHelper {
+    private static readonly _debug = false;
     private _tree: AuxCausalTree;
     private _userId: string;
     private _lib: SandboxLibrary;
@@ -105,7 +106,9 @@ export class FileHelper {
      * @param tags (Optional) The tags that the file should have.
      */
     async createFile(id?: string, tags?: File['tags']): Promise<void> {
-        console.log('[FileManager] Create File');
+        if (FileHelper._debug) {
+            console.log('[FileManager] Create File');
+        }
 
         const file = createFile(id, tags);
         await this._tree.addFile(file);
@@ -118,7 +121,9 @@ export class FileHelper {
         builderContextId?: string,
         contextType?: unknown
     ): Promise<void> {
-        console.log('[FileManager] Create File');
+        if (FileHelper._debug) {
+            console.log('[FileManager] Create File');
+        }
 
         const workspace: Workspace = createWorkspace(
             undefined,
@@ -136,13 +141,22 @@ export class FileHelper {
      * @param arg The argument that should be passed to the event handlers.
      */
     actionEvents(eventName: string, files: File[], arg?: any) {
-        console.log('[FileManager] Run event:', eventName, 'on files:', files);
+        if (FileHelper._debug) {
+            console.log(
+                '[FileManager] Run event:',
+                eventName,
+                'on files:',
+                files
+            );
+        }
 
         // Calculate the events on a single client and then run them in a transaction to make sure the order is right.
         const fileIds = files ? files.map(f => f.id) : null;
         const actionData = action(eventName, fileIds, this._userId, arg);
         const result = calculateActionEvents(this._tree.value, actionData);
-        console.log('  result: ', result);
+        if (FileHelper._debug) {
+            console.log('  result: ', result);
+        }
 
         return result;
     }